--warnings
--warnings will now cause rate limits and warnings to be written to stdout
along with the expected information (tweets, users, etc).

the reason for this is that these rate limits and warnings could be pretty
important for understanding the data later. For example in cases when you are
collecting from a high volume event you may only be getting a portion of the
tweets which is a significant property of the dataset for someone working with
it later.

fixes #121