Replies
This commit includes functionality and a test for getting replies for a given
tweet. Twarc.replies operates on an iterator of tweet objects, since a user
screen name and the tweet_id are needed. The filename can be passed in
on the command line.

    twarc replies tweets.json > replies.json

Since Twitter's API doesn't support getting replies to a given tweet
twarc will use the search api to look for responses to a given user and then
filter those for tweets where the in_reply_to_status_id_str matches the
original tweet id. It also will recursively fetch replies to replies, unless
told not to.