config loading
This commit includes some changes to the way that the twitter keys are
loaded, including a new MissingKeys exception for programmatic use of
twarc. By default the twarc.Twarc constructor will now look for a config
file at the default location and use it if it is there. This should
simplify the writing of programs that use twarc so they don't have to
go fishing around for the keys in the environment and config files themselves.

fixes #147
fixes #150
fixes #132
fixes #152