csv output
This commmit pulls utils/json2csv.py into twarc proper. This means you can
now tell twarc to output directly to csv if you aren't interested in the
json.

    twarc search obama --format json > obama.csv

The big caveat here is that information is lost when you convert to
csv. But you might not be in a situation where you care about that.

You can also tell twarc to output to a filename:

    twarc search obama --output obama.json

And, optionally to split the output using a certain number of lines:

    twarc search obama --output obama.json --split 1000

Which will result in numbered files being written to the filesystem:

    obama-001.json
    obama-002.json
    obama-003.json
    ...

Fixes #172