entry_points
Previously on Windows setup.py would install bin/twarc.py instead of
bin/twarc into the system path. Unfortunately this caused twarc.py to
import itself, which caused a failure when it tried to execute the main()
function, which wasn't there.

This commit introduces the use of entry_points in setup.py which
hopefully will handle doing the right thing on Windows and Unix
platforms.

Fixes #194