filter protected on search and filter
This commit adds a safeguard that search and filter do not return
protected tweets unless they were specifically requested. It appears at
the moment that the search API doesn't return protected tweets even if
you are entitled to see them. Adding this protective measure in twarc is
useful if this undocumented behavior changes for whatever reason.

Fixes #219