Allow sample/filter to optionally return keepalive events
The keep-alive signals returned by the Twitter API streaming
endpoints are useful data in and of themselves. Recording them
allows us to differentiate between quiet periods where no tweets
occurred, and periods where the connection to the API was down.

The format for the keep alive events is just a string message
'keep-alive', and should be easy to consume in downstream apps.
The feature is off by default.