Handle 401 in the rate_limit decorator
It's a bit ugly, but catches and reraises the requests HTTPError
on a 401 Authentication Required response, but augments it with
a slightly more friendly error message for the user.

Also adds a flag to turn off validation on keys, mostly for tests
but might be need for CLI interface.

fixes #227