remove from_date/to_date
This commit removes the the from_date and to_date parameters from the
search() method since the Twitter standard search API does not use them.

This commit also fixes a problem introduced in
1e0b5df3c232983e9c1bd45123968a41eb591e67 which caused max_id not to get
updated properly while retrieving multiple pages of results. This caused
the since_id to be greater than max_id, which in turn caused the Twitter
API to throw an HTTP 403 error.