Log when terminated
The twarc command catches SIGTERM so that users can ctrl-c to stop the
process. But it did so quietly, which could be perplexing when users put
twarc into the background and then logged out (which can cause SIGTERM
being sent to the process). To aid in diagnosing when/why things stop
this commit adds logging to when twarc is shutting down. fixes #319