Provide getting started message to client on first run.
Provide message to client on getting started when credentials
have not yet been supplied. Previously, a stack trace was shown
attributing the exception/error that was invoked. This fix
makes the output when getting started more user friendly and
gives the user an action to perform to progress their usage of
the tool.

Closes #330