v2
This commit wipes the deck clean for work on a v2 version of twarc that
will work with the v2 Twitter API. This is an opportunity to take what
was learned from v1 and improve on it, while letting some of the cruft
that is in v1 go. We can move things over as needed from the main
branch.