main function
This commit moves the majority of the processing code into a main
function which is only called when the program is run from the command
line and not when it is invoked as part of the multiprocessing queue.
This is to hopefully fix an error encountered on OS X with Python 3.9.