Barebones Twarc2 client object
Implements the basic structure of the Twarc2 object. Currently
only the bearer_token application authentication method is
supported, and the bearer_token must be passed into the object.

Authentication is handled by injecting the bearer_token as an
authorisation header into a requests session - no OAuth
infrastructure is required.

This also adds the expansions enumerated in the wiki into the code
itself, so we can grow them there/use them in other places when
we figure out if we need to do anything else in particular for
customisation.