Flatten
This commit adds a flatten parameter to Twarc2.sample which will use the
twarc.expansions.flatten function to inline all the included extensions.
I think it also fixes a bug in the "geo" expansion that was expecting a
list where there was a dictionary.

The additional test_sample_flattened will use sample with flatten=True
to watch up to 5000 tweets. If it finds all the expected expansions it
will stop. If it hits 5000 without seeing them all it will stop and then
will report which expansion failed to be recognized. This 5000 limit may
need to be adjusted in practice. Things like "geo" don't show up that
often.

Some of the over-protective exception handling (catching Exception) in
Twarc2 was removed for now to help with debugging and seeing where and
when things go wrong.