click
This adds the start of a new command that uses the click module to
manage the twarc command and its various subcommands. At the moment only
sample is supported. You should be able to run the v1.1 API sample like so:

    twarc sample > sample.jsonl

And the v2 currently needs a bearer-token, but that should change in the
future if the bearer-token is fetched automatically given the app keys:

    twarc --v2 sample --bearer-token XXX > sample.jsonl

You can also get the "flattened" output like so:

    twarc --v2 sample --bearer-token XXX --flatten > sample.jsonl

I think the configuration initialization and reading needs to be pulled
out of twarc.client into a new module twarc.config so it can also be
used by twarc.client2?