config
This commit adds support for configuring twarc2 using the
click-config-file plugin. When starting twarc2 if a bearer token cannot
befound the user is prompted to enter a bearer token, which is persisted
to an os-specific location (e.g. ~/.config/twarc/config on linux).

You can still pass a bearer-token explicitly with --bearer-token and
configure via an environment varibale BEARER_TOKEN. The resolution order
is CLI > Environment > Configuration file > Default.

fixes #400