User Authentication
Since we anticipate needing user authentication this commit introduces a
twarc.handshake module that contains a function that interactively asks
the user for their app's consumer_key and consumer_secret in order to generate
iboth a bearer_token and access token and access_token_secrets. It is very
much like the twarc1 interaction.