User Auth client
This commit modifies the Twarc2 constructor so that it takes
consumer_key, consumer_secret, access_token and access_token_secret. It
no longer takes a bearer_token because that can be obtained on the fly.
The Twarc2 client creates two HTTP clients internally, one for user auth
and the other for app auth. At the moment only the app_auth client is
being used. But when we implement support for endpoints that have user
auth we can use it then.

fixes #395