--max-results
This commit adds the --max-results option to search which can control
the max_results parameter that is sent to the search/recent and
search/all endpoints. Ordinarily everyone would want to request the
maximum number of tweets per request, so as to maximize the 15 minute
request quota limit. But some occasional 503 errors have been observed
coming from the search/all endpoint and a forum post has suggested that
dialing max_results down to 100 can help.

Refs #449