Fix misleading log messages in timeline and search
All data is processed when the API indicates there are no more
pages to retrieve, not when we hit an empty page of results. A
page of results can be empty when tweets are deleted.

As an addendum, also fix a subtle problem that might lead to
exceeding the 1 request/s rate limit for search when there are
lots of empty pages.