Remove --flatten and add ensure_flattened()
This commit fixes #473 by removing the `--flatten` command line
option for twarc2. It also adjusts the behavior of
twarc.expansions.flatten to always return a list of dictionaries where
each dictionary is a tweet. Finally it adds
twarc.expansions.ensure_flattened to make sure passed in data is
flattened, which will throw a ValueError if what is passed in does not
look like twitter data. This function is also available as
twarc.ensure_flattened because we expect plugins to use it and don't
want to bury it.