Improved config logging
This commit adds twarc.config.ConfigProvider which is based on
click_config_file.configobj_provider and stores the file path for the
config file that was used. This is useful for logging.

Also when --verbose is used the log will now contain the keys that are
being used to talk to the API. This isn't something you would normally
want in your logs, but it can be useful for debugging situations like #441
and #469.