Catch ChunkedEncodingError in stream as well as sample
This is a refactor of stream and sample so that ChunkedEncodingError is
caught in both. Also checking for disconnect errors is done in both now
instead of only sample. Logging was enhanced to include the ids for
tweets that have been archived, as it was in twarc1. See #471 (again)
for background details about ChunkedEncodingError.