Disable stream under GitHub Actions
For some reason the test_stream test throws an error when running under
GitHub Actions. This is because a 400 error is coming back from Twitter.
I picked the URL out of the log on GitHub and tried it with the
BEARER_TOKEN that's being used and it works fine locally. For the time
being I'm going to disable it so that we don't miss out on the benefit
of having the other tests run on GitHub.