Refactor of exception handling
While trying to address #471 it became increasingly clear that our
approach to exception handling in twarc2 needed some simplification. We
decided to reuse decorators from twarc1 in twarc2 in the hopes that
lessons learned about the v1 API would be useful with v2. Butt some of
the edge cases involving the Premium API in twarc1 are not needed in
twarc2, and the behavior of the twarc v2 endpoint seems to have
significantly different failure modes than the v1.1 endpoint.

To address this twarc.decorators just prior to twarc v2 release has been
restored and a new twarc.decorators2 has been introduced which allows
these exceptions to be developed separately. When twarc v1.1 goes away
we can remove twarc.decorators as we will twarc.client and
twarc.command.

A big change in twarc.decorators2 is that rather than adding different
decorators for different exception types there is now
catch_request_exceptions which will catch *any*
requests.exception.RequestException that is not an
requests.exception.HTTPError. It will log the error and then do an
exponential sleep backoff, reconnect and try again until the maximum
number of tries has been reached (default 30).

HTTP errors are not caught because those need to be handled separately
by other decorators like rate_limit decorator and cli_api_error.