Twarc2._stream
This commit adds an internal method Twarc2._stream which is used by both
the Twarc2.stream and Twarc2.sample methods to connect to the Twitter
API and stream results. It is wrapped with the catch_request_exceptions
decorator, which ought to catch an request errors, log them and retry,
and hopefully this will finally address #505. 🤞