Inlined stream exception handling
The catch_request_exceptions decorator wasn't working well with
streaming responses which return an object successfully, but later throw
an exception when iter_lines() is called and iterated over. Now there is
a somewhat duplicative bit of logic in Twarc2._stream for catching
request exceptions, trying again with an exponential backoff. Refs #505