Check usernames/ids in timelines command
This should address situations where the input usernames to the
timelines command contains things that would cause the API to throw a
400 error, which gets exposed to the user in an unfriendly way.

In the process of doing this I noticed that the timelines command was
unable to process lines containing user ids as advertised. I think this
happened when it was adjusted to allow quoted strings in the input
source.