Move last-30-days by default fix from the CLI to the client
This fixes bugs in timelines and conversations, which use the
search_all endpoint without the fix, therefore limiting them to
the last 30 days only.

This fix just moves the default start_time set in the CLI app to
the client app, so it will apply by default to all uses of the
search_all method.