requirements files
I noticed that the readthedocs was failing because humanize was missing
from requirements-mkdocs.txt. Rather than needing to remember to
duplicate things across these files I thought I'd try and pass a list of
requirements files to see if that worked. I also took the opportunity to
alphabetize the requirements files.