Catch additional edge case in the counts API
A search for only a user that doesn't exist like "from:10", doesn't
return any data, just immediately creates a response with a meta
key indicating there are zero matching tweets. This conflicts with
the workaround to catch early termination of counts, and was leading
to errors being raised.

This change now differentiates between the two failure modes,
by checking if any rows of data have been returned previously,
or if this is an immediate early termination like happens when
a user doesn't exist anymore.